pyproject.toml
[tool.poetry]
name = "base.django"
version = "0.1.0"
description = ""
authors = ["Trung Nguyen <nttrung143@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
amqp = "^5.0.2"
billiard = "^3.6.3"
celery = "^5.0.2"
dj-database-url = "^0.5.0"
django-filter = "^2.4.0"
djangorestframework = "^3.12.2"
flower = "^0.9.5"
gunicorn = "^20.0.4"
kombu = "^5.0.2"
numpy = "^1.19.4"
olefile = "^0.46"
pandas = "^1.1.4"
psycopg2-binary = "^2.8.6"
python-dateutil = "^2.8.1"
pytz = "^2020.4"
redis = "^3.5.3"
rest-pandas = "^1.1.0"
six = "^1.15.0"
vine = "^5.0.0"
whitenoise = "^5.2.0"
Babel = "^2.8.0"
Django = "^3.1.3"
Pillow = "^8.0.1"
python-dotenv = "^0.15.0"
django-debug-toolbar = "^3.1.1"
watchdog = "^0.10.3"
tornado = "^6.1"
graphene-django = "^2.13.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
